requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
alembic==1.12.1
google-cloud-vision==3.4.5
openai==1.3.0
python-multipart==0.0.6
Pillow==10.1.0
python-jose==3.3.0
passlib==1.7.4
stripe==7.3.0
httpx==0.25.1
pytest==7.4.3